UIORGS-390 Move reusable constants and hooks to appropriate places
