--- src/Settings/constants.js
@@ -1,8 +1,0 @@
-export const SETTINGS_API = 'organizations-storage/settings';
-export const BANKING_ACCOUNT_TYPES_API = 'organizations-storage/banking-account-types';
-
-export const BANKING_INFORMATION_ENABLED_QUERY_KEY = 'BANKING_INFORMATION_ENABLED';
-export const BANKING_INFORMATION_SEARCH_PARAMS = {
-  query: `key=${BANKING_INFORMATION_ENABLED_QUERY_KEY}`,
-  limit: 1,
-};
