+++ src/common/constants/api.js
@@ -1,7 +1,9 @@
+export const AGREEMENTS_API = 'erm/sas';
+export const BANKING_ACCOUNT_TYPES_API = 'organizations-storage/banking-account-types';
 export const CATEGORIES_API = 'organizations-storage/categories';
 export const CONTACTS_API = 'organizations-storage/contacts';
 export const INTERFACES_API = 'organizations-storage/interfaces';
+export const SETTINGS_API = 'organizations-storage/settings';
 export const TYPES_API = 'organizations-storage/organization-types';
-export const AGREEMENTS_API = 'erm/sas';
 
 export const MAX_LIMIT = 2147483647;
