+++ src/common/constants/organization.js
@@ -1,3 +1,4 @@
 export { ORGANIZATION_STATUS } from '@folio/plugin-find-organization';
 
+export const BANKING_INFORMATION_ENABLED_KEY = 'BANKING_INFORMATION_ENABLED';
 export const DICT_CATEGORIES = 'vendorCategories';
