+++ src/common/hooks/index.js
@@ -1,4 +1,5 @@
 export * from './useAcqMethods';
+export * from './useBankingInformation';
 export * from './useIntegrationConfig';
 export * from './useIntegrationConfigMutation';
 export * from './useLinkedAgreements';
